+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Increment `@folio/stripes` to `v5`, `react-router` to `v5.2`.
 * [UITEN-76](https://issues.folio.org/browse/UITEN-76) Refactor forms to use final-form
 * [UITEN-107](https://issues.folio.org/browse/UITEN-107) Retrieve data for up to 1000 plugins
+* [UITEN-108](https://issues.folio.org/browse/UITEN-108) Import support locales from stripes-core. 
 
 ## [4.0.0](https://github.com/folio-org/ui-organization/tree/v4.0.0) (2020-06-11)
 [Full Changelog](https://github.com/folio-org/ui-organization/compare/v3.0.0...v4.0.0)